Add support for apachepulsar/pulsar-all image
`apachepulsar/pulsar-all` brings connectors in the image.
