+++ modules/pulsar/src/main/java/org/testcontainers/containers/PulsarContainer.java
@@ -7,7 +7,7 @@ import org.testcontainers.utility.DockerImageName;
 /**
  * Testcontainers implementation for Apache Pulsar.
  * <p>
- * Supported image: {@code apachepulsar/pulsar}
+ * Supported images: {@code apachepulsar/pulsar}, {@code apachepulsar/pulsar-all}
  * <p>
  * Exposed ports:
  * <ul>
@@ -64,7 +64,7 @@ public class PulsarContainer extends GenericContainer<PulsarContainer> {
 
     public PulsarContainer(final DockerImageName dockerImageName) {
         super(dockerImageName);
-        dockerImageName.assertCompatibleWith(DockerImageName.parse("apachepulsar/pulsar"));
+        dockerImageName.assertCompatibleWith(DEFAULT_IMAGE_NAME, DockerImageName.parse("apachepulsar/pulsar-all"));
         withExposedPorts(BROKER_PORT, BROKER_HTTP_PORT);
         setWaitStrategy(waitAllStrategy);
     }
